Add files via upload
The purpose of this script is to configure the network profiles on a macbook laptop, for multiple sites (docking stations) that have their own respective static network configurations. A launchd daemon is set to run in the background on the macbook, watching for the usb connection or disconnection identified by the docking station mac address. When a connection is established it will create, populate and save the respective network configuration, if the profile exists, the configuration will be verified and corrected if needed. When a disconnection is detected the network profile is switch back to the "Automatic" profile, allowing the user to scan and connect to wifi as usual.